Extract JS module for checksum calculation
Refactoring in preparation for changing the checksum code so that
it calculates the hash incrementally rather than loading the entire
file into memory.
